Home, sheet, room pages and their operations are all inaccessible for a unauthenticated user. Users can only see their own sheets unless they are a superuser, then they can see all sheets. Additional formatting in HTML and javaScript was added to calculate changes of modifiers for character sheets and for better organization. Refactoring of the CRUD operation method names in view.py. Added new documentation throughout the project. Sheet now automatically adds the user to its field, users no longer can assign a user to a sheet (unless in the admin panel).
Spike Epic #1 should now be complete with the addition of storing features and spells in a textfield. As a user can now create a sheet, add statistics, options and notes to the character and save it to their account. As stated by US1. The user can also now modify statistics of the character, remove or insert new character options/notes and remove the character sheet from their account as stated by US2.

Fixed an issue where page would 404 after sheet creation even though sheet was properly created as expected.

Signed-off-by: <jesse.selin@sasktel.net>
Signed-off-by: Senpai-Chans Yaoi PC <jesse.selin@sasktel.net>